Feat/update state manager (#778)
* feat(TeamStateManager): add support for WASM and Server modes

- Implemented ITeamStateManager interface for managing team states.
- Added registration logic for team state manager in ServiceCollectionExtensions.cs.
- Created ServerTeamStateManager and WasmTeamStateManager to handle team state management in respective environments.
- Updated User.razor to utilize the new team state manager for fetching user team information.

* feat:update WasmTeamStateManager refresh token

* fix:navigate page

* fix(WasmTeamStateManager): update token refresh logging

Updated the log message for token refresh failure and removed the null check for newToken. The logic now directly navigates to the page refresh on token refresh failure, simplifying the code.

* feat: Remove team state retrieval methods

Removed the methods for retrieving and clearing the current team ID from the ITeamStateManager interface and its implementations in ServerTeamStateManager and WasmTeamStateManager. The User.razor file was updated to directly use MasaUser.CurrentTeamId instead of relying on the authentication state for the current team ID.